requirements.txt
PySide2
fake-bpy-module-2.82
pypiwin32
